+++ docs/RELEASES.md
@@ -16,7 +16,7 @@ Legend:
 
 ## <a name="6.0.1.0"></a>6.0.1.0 :sheep: <sup>`3`</sup>
 - [ ] Robot to delete not accessed users [#49]
-- [ ] `210421>......` Run robot each hour and remove it from request
+- [x] `210421>210430` Run robot each 15 minutes and remove it from request
 - [x] `210421>210421` Fix problem closing account (one of the locals is not closing)
 
 ## <a name="6.0.0.0"></a>6.0.0.0 :dragon: <sup>`7`</sup>